Refactor almost everything for routing types
Refactor all the code so that the type of routing (default fastest vs.
traffic-based) is defined at the start as a command-line argument, and the
subsequent optimizations and code paths use this. Comment out a LOT of
extra code that was unnecessary (either entirely so, because it calculated
things we don't care about, or because the refactor rendered it obsolete),
and the next commit should probably remove these.